Move `WebRemoteDynamicClaimsPrincipalContributor` to `AbpAspNetCoreAuthenticationJwtBearerModule`.
